+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Use GitHub Actions for CI and pkgdown site
 * Add project options to initialize with git repository and/or renv private library
+* Add conditions in `use_references()` to skip modifications if they already exist
 
 # rsf 0.2.2